lib/requirements.txt
###### Python libraries ######

numpy==1.18.1
tensorflow==2.1.0
wave==0.0.2
pydub==0.23.1
pyyaml==5.3
scipy==1.4.1
tqdm==4.41.1
pyaudio==0.2.11
librosa==0.7.2
rwave==0.1.1
